Rename IcdManagement to ICDManagement for consistency.
