+++ src/app/clusters/icd-management-server/app_config_dependent_sources.cmake
@@ -16,8 +16,8 @@
 TARGET_SOURCES(
   ${APP_TARGET}
   PRIVATE
-    "${CLUSTER_DIR}/IcdManagementCluster.cpp"
-    "${CLUSTER_DIR}/IcdManagementCluster.h"
+    "${CLUSTER_DIR}/ICDManagementCluster.cpp"
+    "${CLUSTER_DIR}/ICDManagementCluster.h"
 )
 
 # TODO(#32321): Remove after issue is resolved
