+++ src/app/clusters/icd-management-server/app_config_dependent_sources.gni
@@ -12,6 +12,6 @@
 # See the License for the specific language governing permissions and
 # limitations under the License.
 app_config_dependent_sources = [
-  "IcdManagementCluster.cpp",
-  "IcdManagementCluster.h",
+  "ICDManagementCluster.cpp",
+  "ICDManagementCluster.h",
 ]
